requirements.txt
pandas
requests
beautifulsoup4
lyricsgenius
psycopg2-binary
numpy
